Create a package for common helpers used accros packages and move "retrier" and "promises" there
