+++ packages/framework-common-helpers/.mocharc.yml
@@ -1,0 +1,10 @@
+diff: true
+require: 'ts-node/register'
+extension:
+  - ts
+package: './package.json'
+recursive: true
+reporter: 'spec'
+timeout: 5000
+full-trace: true
+bail: true
