+++ packages/framework-common-helpers/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@boostercloud/framework-types",
+  "name": "@boostercloud/framework-common-helpers",
   "version": "0.14.1",
-  "description": "Contains Booster types related to the information extracted from the user project",
+  "description": "Contains Booster common helpers used by the core and provider packages",
   "keywords": [
-    "framework-types"
+    "framework-common-helpers"
   ],
   "author": "Booster Cloud",
   "homepage": "https://booster.cloud",
@@ -32,14 +32,12 @@
     "url": "https://github.com/boostercloud/booster/issues"
   },
   "dependencies": {
-    "@types/graphql": "14.5.0",
-    "uuid": "8.3.2"
+    "@boostercloud/framework-types": "0.14.1"
   },
   "devDependencies": {
     "@types/uuid": "8.3.0",
     "chai": "4.2.0",
     "chai-as-promised": "7.1.1",
-    "fast-check": "^1.18.1",
     "sinon-chai": "3.5.0"
   }
 }
