+++ packages/framework-common-helpers/src/test-helpers/expect.ts
@@ -1,0 +1,6 @@
+import * as chai from 'chai'
+
+chai.use(require('sinon-chai'))
+chai.use(require('chai-as-promised'))
+
+export const expect = chai.expect
