+++ packages/framework-core/package.json
@@ -29,7 +29,8 @@
     "url": "https://github.com/boostercloud/booster/issues"
   },
   "dependencies": {
-    "@boostercloud/framework-types": "^0.14.1",
+    "@boostercloud/framework-types": "0.14.1",
+    "@boostercloud/framework-common-helpers": "0.14.1",
     "fp-ts": "^2.0.3",
     "graphql": "15.4.0",
     "graphql-subscriptions": "1.1.0",
