+++ packages/framework-provider-kubernetes-infrastructure/package.json
@@ -20,6 +20,7 @@
     "url": "git+https://github.com/boostercloud/booster.git"
   },
   "dependencies": {
+    "@boostercloud/framework-common-helpers": "0.14.1",
     "@kubernetes/client-node": "^0.12.0",
     "archiver": "4.0.1",
     "body-parser": "^1.18.3",
